Added lists

+++ src/main/java/com/emerycprimeau/transfer/GameRequest.java
@@ -1,5 +1,5 @@
 package com.emerycprimeau.transfer;
 
 public class GameRequest {
-    public Long userId;
+    public int userId;
 }
